+++ src/helpers/index.js
@@ -1,11 +1,7 @@
 const createError = require("./createError");
 const ctrlWrapper = require("./ctrlWrapper");
-const sendEmail = require("./sendEmail");
-const sendVerifyEmail = require("./sendVerifyEmail");
 
 module.exports = {
   createError,
   ctrlWrapper,
-  sendEmail,
-  sendVerifyEmail,
 };
