+++ src/services/contacts/updateContactById.js
@@ -1,7 +1,10 @@
 const { Contact } = require("../../models/contact");
 
 const updateContactById = (contactId, contact) => {
-  const result = Contact.findByIdAndUpdate(contactId, contact, { new: true });
+  const result = Contact.findByIdAndUpdate(contactId, contact, {
+    select: "-createdAt -updatedAt",
+    new: true,
+  });
   return result;
 };
