+++ src/services/contacts/updateContactFavoriteById.js
@@ -1,7 +1,10 @@
 const { Contact } = require("../../models/contact");
 
 const updateContactFavoriteById = (contactId, value) => {
-  const result = Contact.findByIdAndUpdate(contactId, value, { new: true });
+  const result = Contact.findByIdAndUpdate(contactId, value, {
+    select: "-createdAt -updatedAt",
+    new: true,
+  });
   return result;
 };
