+++ src/services/email/index.js
@@ -1,0 +1,7 @@
+const sendEmail = require("./sendEmail");
+const sendVerifyEmail = require("./sendVerifyEmail");
+
+module.exports = {
+  sendEmail,
+  sendVerifyEmail,
+};
